Add .vim/swap directory
This is the swap directory I use. If this directory doesn't exist, vim
will complain---loudly. So, I want to put this in the dotfiles repo. But
because git does not support adding directories (only files), it will
not include .vim/swap without files in it (even though I don't actually
want to track any files inside of it). Therefore I added an empty
.gitignore file to .vim/swap, allowing the directory to be included as
a by-product.