Separate functions from filetype configurations
Loading multiple files of the same filetype reloads all functions
defined in the ftyplugin/<filetype>.vim file, raising errors if the
functions are already defined (and if the functions aren't defined with
a `!` after their name). This is not ideal, so I've moved all custom
functions to the autoload/ftplugin/<filetype>.vim files, where they will
only load when a file of the right time is loaded and where they will
only load once.