Add screen terminal to case statements
Manually setting the $TERM variable in .*rc files is not a good idea, so
I recmoved if a few commits back. This caused window title setting in
tmux to fail because tmux sets the terminal to screen-256color instead
of xterm-256color. So, screen is added to the relevant case statements.